requirements.txt
matplotlib==3.10.3
numpy==2.1.1
tqdm==4.67.0
Pillow==10.3.0